Revise action argparse defaults and error messages
Change the program name to 'action' and preprocess sys.argv arguments so the Python script name is no longer a part of the error messages.

Remove required arguments and have default values for all arguments as per Lan Xia's (@llxia) suggestion:

>>>
I think we should have the default values for the following parameters:
--sdk_resource = nightly
--build_list = openjdk
--target = jdk_math
--platform = x86-64_linux
--jdk_version = 8
--jdk_impl = openj9

We can adjust this later to cover the common use case.